+++ lib/variable/operations_common.h
@@ -12,21 +12,10 @@
 namespace scipp::variable {
 
 // Helpers for in-place reductions and reductions with groupby.
-SCIPP_VARIABLE_EXPORT void sum_impl(Variable &summed, const Variable &var);
-SCIPP_VARIABLE_EXPORT void all_impl(Variable &out, const Variable &var);
-SCIPP_VARIABLE_EXPORT void any_impl(Variable &out, const Variable &var);
-SCIPP_VARIABLE_EXPORT void max_impl(Variable &out, const Variable &var);
-SCIPP_VARIABLE_EXPORT void min_impl(Variable &out, const Variable &var);
 SCIPP_VARIABLE_EXPORT Variable mean_impl(const Variable &var, const Dim dim,
                                          const Variable &masks_sum);
-SCIPP_VARIABLE_EXPORT Variable &mean_impl(const Variable &var, const Dim dim,
-                                          const Variable &masks_sum,
-                                          Variable &out);
 SCIPP_VARIABLE_EXPORT Variable nanmean_impl(const Variable &var, const Dim dim,
                                             const Variable &masks_sum);
-SCIPP_VARIABLE_EXPORT Variable &nanmean_impl(const Variable &var, const Dim dim,
-                                             const Variable &masks_sum,
-                                             Variable &out);
 
 template <class T> T normalize_impl(const T &numerator, T denominator) {
   // Numerator may be an int or a Eigen::Vector3d => use double
@@ -36,12 +25,6 @@ template <class T> T normalize_impl(const T &numerator, T denominator) {
   denominator.setUnit(units::one);
   return numerator *
          reciprocal(astype(denominator, type, CopyPolicy::TryAvoid));
-}
-
-template <class T> void normalize_inplace_impl(T &numerator, T denominator) {
-  denominator.setUnit(units::one);
-  numerator *= reciprocal(
-      astype(denominator, core::dtype<double>, CopyPolicy::TryAvoid));
 }
 
 SCIPP_VARIABLE_EXPORT void expect_valid_bin_indices(const Variable &indices,